Manually merged back the original main.rs to aid cherry picking, with some luck.
